created incidents dir; separated add dialog to another file; updated imports
