+++ src/components/medicals/incidents/Incident.js
@@ -1,6 +1,5 @@
 import React, { useState, useEffect } from "react";
-import useSimpleAuth from "../../hooks/ui/useSimpleAuth";
-import MedicalRepository from "../../repositories/MedicalRepository";
+import MedicalRepository from "../../../repositories/MedicalRepository";
 
 import Accordion from "@mui/material/Accordion";
 import AccordionDetails from "@mui/material/AccordionDetails";
