+++ src/components/routes/MedicalRoutes.js
@@ -2,7 +2,7 @@ import React, { useEffect, useState } from "react";
 import { Route } from "react-router-dom";
 import { ViewMedical } from "../medicals/ViewMedical";
 import { PetMedicalBio } from "../medicals/bio/PetMedicalBio"
-import { IncidentsList } from "../medicals/IncidentsList"
+import { IncidentsList } from "../medicals/incidents/IncidentsList"
 import { VetVisitsList } from "../medicals/VetVisitsList";
 import { VaccinationsList } from "../medicals/VaccinationsList";
 import { MedicationsList } from "../medicals/MedicationsList";
